requirements.txt
streamlit==1.28.0
supabase==2.0.0
python-dotenv==1.0.0
streamlit-extras==0.3.5
Pillow==10.0.0
pandas==2.1.0
numpy==1.24.0
